Move SettingsViewModel to own file
